make the userop page default
