+++ examples/testapp/src/components/Layout.tsx
@@ -30,6 +30,7 @@ export const WIDTH_2XL = '1536px';
 
 const PAGES = [
   '/',
+  '/dashboard',
   '/add-sub-account',
   '/import-sub-account',
   '/auto-sub-account',
